New Centroid Functions for K-centroids clustering of (ordinal) categorical/mixed data

Functions to calculate cluster centroids for K-centroids clustering that extend the options available in package flexclust.

centMode calculates centroids based on the mode of each variable. centMin minimizes the applied distance metric to find centroids within a specified range. centOptimNA replicates the exact behaviour of flexclust::centOptim, just with NA removal.

These functions are designed for use within flexclust::kcca or functions that are built upon it. Their use is easiest via the wrapper kccaExtendedFamily, for more information see ?kccaExtendedFamily.

centMode(x)

centMin(x, dist, xrange = NULL)

centOptimNA(x, dist)

Arguments

x

A numeric matrix or data frame. Categorical/ordinal variables need to be coded as 1:length(levels(x[,i])) in steps of one.

dist

The distance measure function to be minimized in centMin.

xrange

Range specification for the variables. Currently only used for centMin. Possible values are:

NULL (default): defaults to 'all'.

'all': Uses the range of the entire data set.

'columnwise': Uses column-specific ranges.

A numeric vector of c(min, max): Applies the specified range to all columns.

A list of numeric vectors of c(min,max): Uses the user-specified columnwise ranges, the length of the list must be equal to the number of columns to be scaled.

Value

A named numeric vector containing the centroid values for each column of x.

Details

centMode: Column-wise modes are used as centroids, and ties are broken randomly. In combination with Simple Matching Distance (distSimMatch), this results in the kmodes algorithm.

centMin: Column-wise centroids are calculated by minimizing the specified distance measure between a value of x, and all possible levels of x.

centOptimNA: Column-wise centroids are calculated by minimizing the specified distance measure via a general purpose optimizer. Unlike in flexclust::centOptim, NAs are removed from the starting search values.

See also

Examples

# Example: Mode as centroid
dat <- data.frame(A = rep(2:5, 2),
                  B = rep(1:4, 2),
                  C = rep(c(1, 2, 4, 5), 2))
centMode(dat)
#> A B C 
#> 2 1 2 
## within kcca
flexclust::kcca(dat, 3, family=kccaExtendedFamily('kModes')) #default centroid
#> kcca object of family ‘kModes’ 
#> 
#> call:
#> flexclust::kcca(x = dat, k = 3, family = kccaExtendedFamily("kModes"))
#> 
#> cluster sizes:
#> 
#> 1 2 3 
#> 2 2 4 
#> 

# Example: Centroid is level for which distance is minimal
centMin(dat, flexclust::distManhattan, xrange = 'all')
#> A B C 
#> 3 2 3 
## within kcca
flexclust::kcca(dat, 3,
                family=flexclust::kccaFamily(dist=flexclust::distManhattan,
                                             cent=\(y) centMin(y, flexclust::distManhattan,
                                                               xrange='all')))
#> kcca object of family ‘flexclust::distManhattan’ 
#> 
#> call:
#> flexclust::kcca(x = dat, k = 3, family = flexclust::kccaFamily(dist = flexclust::distManhattan, 
#>     cent = function(y) centMin(y, flexclust::distManhattan, xrange = "all")))
#> 
#> cluster sizes:
#> 
#> 1 2 
#> 6 2 
#> 
                             
# Example: Centroid calculated by general purpose optimizer with NA removal
nas <- sample(c(TRUE, FALSE), prod(dim(dat)),
              replace=TRUE, prob=c(0.1,0.9)) |> 
       matrix(nrow=nrow(dat))
dat[nas] <- NA
centOptimNA(dat, flexclust::distManhattan)
#> A B C 
#> 4 3 4 
## within kcca
flexclust::kcca(dat, 3, family=kccaExtendedFamily('kGower')) #default centroid
#> Warning: No column-wise distance measures specified, default measures
#>             for each column will be used.
#> kcca object of family ‘kGower’ 
#> 
#> call:
#> flexclust::kcca(x = dat, k = 3, family = kccaExtendedFamily("kGower"))
#> 
#> cluster sizes:
#> 
#> 1 2 3 
#> 2 4 2 
#>